+++ src/main/java/com/project/board/repository/InterBoardRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface InterBoardRepository extends JpaRepository<Board, Long> {
+public interface InterBoardRepository extends JpaRepository<Board, Integer> {
 /*
     Page<Board> search(String searchKeyword, Pageable pageable);*/
